update and fix dashboard

+++ app/Http/Controllers/ReportController.php
@@ -11,16 +11,15 @@ use Yajra\DataTables\Facades\DataTables;
 class ReportController extends Controller
 {
     public $indexof = 1;
+
     public function index(Request $request)
     {
         $query = Report::with('user');
 
-        // Apply filter only for non-admin users
         if (auth()->user()->role_id !== AppHelper::USER_SUPER_ADMIN && auth()->user()->role_id !== AppHelper::USER_ADMIN) {
             $query->where('user_id', auth()->id());
         }
 
-        // Handle AJAX request
         if ($request->ajax()) {
             $reports = $query->get();
 
@@ -30,16 +29,16 @@ class ReportController extends Controller
                     return '<img class="img-responsive center" style="height: 35px; width: 35px; object-fit: cover; border-radius: 50%;" src="' . $photoUrl . '" >';
                 })
                 ->addColumn('id_card', function ($data) {
-                    return $data->user->id_card ?? 'N/A';
+                    return $data->user->staff_id_card ?? 'N/A';
                 })
                 ->addColumn('name', function ($data) {
-                    return $data->user->family_name . $data->user->name ?? 'N/A';
+                    return $data->user->family_name . ' ' . $data->user->name ?? 'N/A';
                 })
                 ->addColumn('area', function ($data) {
                     return __($data->area);
                 })
-                ->addColumn('depot_stock', function ($data) {
-                    return __($data->depot_stock);
+                ->addColumn('outlet', function ($data) {
+                    return __($data->outlet);
                 })
                 ->addColumn('250ml', function ($data) {
                     return __($data->{"250_ml"}) ?? 'N/A';
@@ -53,10 +52,12 @@ class ReportController extends Controller
                 ->addColumn('1500ml', function ($data) {
                     return __($data->{"1500_ml"}) ?? 'N/A';
                 })
+                ->addColumn('location', function ($data) {
+                    return __($data->city .',' .$data->country) ?? 'N/A';
+                })
                 ->addColumn('date', function ($data) {
                     return $data->date ? \Carbon\Carbon::parse($data->date)->format('d-M-Y') : 'N/A';
                 })
-                
                 ->addColumn('other', function ($data) {
                     return __($data->other) ?? 'N/A';
                 })
@@ -69,7 +70,7 @@ class ReportController extends Controller
                                 <i class="fa fa-edit"></i>
                             </a>
                         </span>
-                        &nbsp;&nbsp;
+                          
                         <span class="change-action-item">
                             <a href="' . $deleteRoute . '" class="btn btn-danger btn-sm delete" title="Delete">
                                 <i class="fa fa-fw fa-trash"></i>
@@ -93,41 +94,19 @@ class ReportController extends Controller
     {
         $rules = [
             'area' => 'required',
-            'depot_stock' => 'required',
+            'outlet' => 'required',
             'date' => 'required|date',
             'latitude' => 'required|numeric',
             'longitude' => 'required|numeric',
+            'city' => 'required|string',
+            'country' => 'required|string',
         ];
         $this->validate($request, $rules);
-
-        // Get location details from Google Maps API
-        $apiKey = env('GOOGLE_MAPS_API_KEY');
-        $response = Http::get("https://maps.googleapis.com/maps/api/geocode/json", [
-            'latlng' => $request->latitude . ',' . $request->longitude,
-            'key' => $apiKey
-        ]);
-
-        $city = 'Unknown';
-        $country = 'Unknown';
-
-        if ($response->successful()) {
-            $results = $response->json()['results'];
-            if (!empty($results)) {
-                foreach ($results[0]['address_components'] as $component) {
-                    if (in_array('locality', $component['types'])) {
-                        $city = $component['long_name'];
-                    }
-                    if (in_array('country', $component['types'])) {
-                        $country = $component['long_name'];
-                    }
-                }
-            }
-        }
-
+        // dd($request->all());
         Report::create([
             'user_id' => auth()->id(),
             'area' => $request->area,
-            'depot_stock' => $request->depot_stock,
+            'outlet' => $request->outlet,
             'date' => $request->date,
             '250_ml' => $request['250_ml'],
             '350_ml' => $request['350_ml'],
@@ -136,8 +115,8 @@ class ReportController extends Controller
             'other' => $request->other,
             'latitude' => $request->latitude,
             'longitude' => $request->longitude,
-            'city' => $city,
-            'country' => $country,
+            'city' => $request->city,
+            'country' => $request->country,
         ]);
 
         return redirect()->route('report.index')->with('success', "Report has been created!");
@@ -155,42 +134,24 @@ class ReportController extends Controller
     public function update(Request $request, $id)
     {
         $report = Report::find($id);
+        if (!$report) {
+            return redirect()->route('report.index');
+        }
+
         $rules = [
             'area' => 'required',
-            'depot_stock' => 'required',
+            'outlet' => 'required',
             'date' => 'required|date',
             'latitude' => 'required|numeric',
             'longitude' => 'required|numeric',
+            'city' => 'required|string',
+            'country' => 'required|string',
         ];
         $this->validate($request, $rules);
 
-        // Get location details from Google Maps API
-        $apiKey = env('GOOGLE_MAPS_API_KEY');
-        $response = Http::get("https://maps.googleapis.com/maps/api/geocode/json", [
-            'latlng' => $request->latitude . ',' . $request->longitude,
-            'key' => $apiKey
-        ]);
-
-        $city = 'Unknown';
-        $country = 'Unknown';
-
-        if ($response->successful()) {
-            $results = $response->json()['results'];
-            if (!empty($results)) {
-                foreach ($results[0]['address_components'] as $component) {
-                    if (in_array('locality', $component['types'])) {
-                        $city = $component['long_name'];
-                    }
-                    if (in_array('country', $component['types'])) {
-                        $country = $component['long_name'];
-                    }
-                }
-            }
-        }
-
         $report->update([
             'area' => $request->area,
-            'depot_stock' => $request->depot_stock,
+            'outlet' => $request->outlet,
             'date' => $request->date,
             '250_ml' => $request['250_ml'],
             '350_ml' => $request['350_ml'],
@@ -199,8 +160,8 @@ class ReportController extends Controller
             'other' => $request->other,
             'latitude' => $request->latitude,
             'longitude' => $request->longitude,
-            'city' => $city,
-            'country' => $country,
+            'city' => $request->city,
+            'country' => $request->country,
         ]);
 
         return redirect()->route('report.index')->with('success', "Report has been updated!");
@@ -209,7 +170,10 @@ class ReportController extends Controller
     public function destroy($id)
     {
         $report = Report::find($id);
-        $report->delete();
-        return redirect()->back()->with('success', "Report has been deleted!");
+        if ($report) {
+            $report->delete();
+            return redirect()->back()->with('success', "Report has been deleted!");
+        }
+        return redirect()->back()->with('error', "Report not found!");
     }
 }
